% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmda.R
\name{hmda_filers}
\alias{hmda_filers}
\title{Fetch HMDA filer institutions for a given year}
\usage{
hmda_filers(years)
}
\arguments{
\item{years}{Integer year (e.g. 2022)}
}
\value{
tibble: lei (character), name (character), period (integer)
}
\description{
Returns a list of institutions that filed HMDA data.
}
